[expo] make expo AppEntry work in bare workflow
