+++ packages/expo/src/Expo.fx.ts
@@ -1,0 +1,4 @@
+import './environment/muteWarnings.fx';
+
+// load expo-asset immediately to set a custom `source` transformer in React Native
+import 'expo-asset';
